Refactor GridManager into partial classes
- This class will get giant... preemptively separate into partials
